1. List.Sort uses BiPartition internally. 2. Set can be partitioned in equivalence classes. 3. List.partition has been removed. 4. Partition object has been removed
