+++ core/src/main/scala/dogs/List.scala
@@ -450,9 +450,6 @@ object List extends ListInstances {
     go(n, a, List.empty[A])
   }
 
-
-  implicit def toPartition[A](aList: List[A]): BiPartition[A] = BiPartition(aList)
-
   implicit def toSorted[A]: Sorted[List] = Sorted.quickSort
 }
